+++ src/03_01_dsb.cpp
@@ -1,9 +1,4 @@
-#include <cpp11.hpp>
-#include <string>
-#include <cstring>
-#include <vector>
-
-using namespace cpp11;
+#include "03_dsb.h"
 
 inline bool is_dsb_open(const char *str, int i, int n)
 {
@@ -271,7 +266,7 @@ void extract_operator(const char *str, int &i, int n,
     std::string string_value = "";
     std::string dsb_value = "";
 
-    int n = std::strlen(str);
+    int n = strlen(str);
 
     int i = 0;
     while (i < n)
@@ -378,7 +373,7 @@ void extract_operator(const char *str, int &i, int n,
 
     const char *str = CHAR(STRING_ELT(Rstr, 0));
 
-    int n = std::strlen(str);
+    int n = strlen(str);
 
     writable::list dsb_element;
     std::vector<std::string> operator_vec;
@@ -422,7 +417,7 @@ inline bool is_if_separator(const char *str, int i, int n, bool semicolon = fals
 {
 
     const char *str = CHAR(STRING_ELT(Rstr, 0));
-    int n = std::strlen(str);
+    int n = strlen(str);
 
     writable::list if_elements;
     std::vector<std::string> operator_vec;