+++ src/03_dsb.h
@@ -1,0 +1,33 @@
+/*
+     _       _     
+    | |     | |    
+  __| | ___ | |__  
+ / _` |/ __|| '_ \ 
+| (_| |\__ \| |_) |
+ \__,_||___/|_.__/ 
+
+Original Author: Laurent R. Berge
+Refactored by Mauricio "Pacha" Vargas Sepulveda starting in Jun 2022
+*/
+
+#pragma once
+
+#include <cpp11.hpp>
+#include <cpp11/doubles.hpp>
+#include <vector>
+#include <cstring>
+
+#include <stdint.h>
+
+#ifdef _OPENMP
+#include <omp.h>
+#else
+#define omp_get_thread_num() 0
+#endif
+
+using namespace cpp11;
+using std::vector;
+using std::fabs;
+using std::strlen;
+
+inline bool is_dsb_open(const char *str, int i, int n);